Refactor storage architecture: move index implementations and consolidate settings
- Move in-memory index implementations from knowpro/ to storage/memory/
- Consolidate all settings classes (MessageTextIndexSettings, RelatedTermIndexSettings, SemanticRefIndexSettings) into knowpro/convsettings.py
- Create IKnowledgeExtractor interface to eliminate import cycles
- Update all imports throughout codebase to use new locations
- Remove obsolete index files from knowpro/ (git detected as renames)
- All type checks passing with pyright

This provides better separation of concerns with storage-specific implementations in storage/ and shared configuration in knowpro/convsettings.py
