+++ python/ta/test/test_property_index_population.py
@@ -15,8 +15,8 @@ from typeagent.aitools.embeddings import AsyncEmbeddingModel
 from typeagent.aitools.utils import load_dotenv
 from typeagent.knowpro.interfaces import Tag, SemanticRef, TextRange, TextLocation
 from typeagent.knowpro import kplib
-from typeagent.knowpro.messageindex import MessageTextIndexSettings
-from typeagent.knowpro.reltermsindex import RelatedTermIndexSettings
+from typeagent.knowpro.convsettings import MessageTextIndexSettings
+from typeagent.knowpro.convsettings import RelatedTermIndexSettings
 from typeagent.aitools.vectorbase import TextEmbeddingIndexSettings
 from typeagent.podcasts.podcast import PodcastMessage
 from typeagent.storage import SqliteStorageProvider
@@ -110,7 +110,7 @@ async def test_property_index_population_from_database():
         # Create a test conversation and build property index
         from typeagent.podcasts.podcast import Podcast
         from typeagent.knowpro.convsettings import ConversationSettings
-        from typeagent.knowpro.propindex import build_property_index
+        from typeagent.storage.memory.propindex import build_property_index
 
         settings2 = ConversationSettings()
         settings2.storage_provider = storage2
