+++ python/ta/test/test_secindex.py
@@ -13,9 +13,9 @@ from fixtures import (
 from typeagent.aitools.embeddings import AsyncEmbeddingModel, TEST_MODEL_NAME
 from typeagent.aitools.vectorbase import TextEmbeddingIndexSettings
 from typeagent.knowpro.convsettings import ConversationSettings
-from typeagent.knowpro.messageindex import MessageTextIndexSettings
-from typeagent.knowpro.reltermsindex import RelatedTermIndexSettings
-from typeagent.knowpro.timestampindex import TimestampToTextRangeIndex
+from typeagent.knowpro.convsettings import MessageTextIndexSettings
+from typeagent.knowpro.convsettings import RelatedTermIndexSettings
+from typeagent.storage.memory.timestampindex import TimestampToTextRangeIndex
 from typeagent.storage.memory import MemoryStorageProvider
 from typeagent.knowpro.secindex import (
     ConversationSecondaryIndexes,
