+++ python/ta/typeagent/knowpro/convknowledge.py
@@ -9,6 +9,7 @@ import typechat
 
 from ..aitools import auth
 from . import kplib
+from .interfaces import IKnowledgeExtractor
 
 
 # TODO: Move ModelWrapper and create_typechat_model() to aitools package.
