+++ python/ta/typeagent/knowpro/query.py
@@ -46,8 +46,8 @@ from .interfaces import (
     Thread,
 )
 from .kplib import ConcreteEntity
-from .messageindex import IMessageTextEmbeddingIndex
-from .propindex import PropertyNames, lookup_property_in_property_index
+from ..storage.memory.messageindex import IMessageTextEmbeddingIndex
+from ..storage.memory.propindex import PropertyNames, lookup_property_in_property_index
 from .searchlib import create_property_search_term, create_tag_search_term_group
