+++ python/ta/typeagent/storage/memory/semrefindex.py
@@ -4,14 +4,13 @@
 from __future__ import annotations
 
 from collections.abc import AsyncIterable, Callable
-from dataclasses import dataclass
 
 from typechat import Failure
 
-from . import convknowledge, kplib, secindex
-from .convknowledge import KnowledgeExtractor
-from .convsettings import ConversationSettings
-from .interfaces import (
+from ...knowpro import convknowledge, kplib, secindex
+from ...knowpro.convsettings import ConversationSettings
+from ...knowpro.convsettings import SemanticRefIndexSettings
+from ...knowpro.interfaces import (
     # Interfaces.
     IConversation,
     IMessage,
@@ -30,15 +29,8 @@ from .interfaces import (
     TextRange,
     Topic,
 )
-from .utils import text_range_from_message_chunk
-from .knowledge import extract_knowledge_from_text_batch
-
-
-@dataclass
-class SemanticRefIndexSettings:
-    batch_size: int
-    auto_extract_knowledge: bool
-    knowledge_extractor: KnowledgeExtractor | None = None
+from ...knowpro.utils import text_range_from_message_chunk
+from ...knowpro.knowledge import extract_knowledge_from_text_batch
 
 
 def text_range_from_location(
